feat(debug): support SSL configuration
gravitee-io/issues#6072
