+++ gravitee-gateway-standalone/gravitee-gateway-standalone-container/src/main/java/io/gravitee/gateway/standalone/vertx/ReactorVerticle.java
@@ -18,6 +18,7 @@ package io.gravitee.gateway.standalone.vertx;
 import io.gravitee.common.http.IdGenerator;
 import io.gravitee.common.utils.Hex;
 import io.gravitee.common.utils.UUID;
+import io.gravitee.gateway.http.vertx.VertxHttpServerConfiguration;
 import io.gravitee.gateway.reactor.Reactor;
 import io.gravitee.gateway.standalone.vertx.ws.VertxWebSocketReactorHandler;
 import io.vertx.core.AbstractVerticle;
